refactor(M2): delete dead computeEconomics (Wave 2 stage 1)
computeEconomics in lib/economics/index.ts had zero live callers — all
production paths use buildMatchEconomics (matching/tce-calculator.ts).

Pre-removal grep output proving zero live callers:
  grep -rn computeEconomics lib/ app/ scripts/ components/ (non-test): 0 matches
  grep -rn economics/index lib/ app/ scripts/ components/ (non-test): 0 matches
  (lib/pipeline.ts had import type only → moved EconomicsInput to lib/types.ts)

Changes:
- delete lib/economics/index.ts
- delete lib/economics/__tests__/index.test.ts
- delete lib/economics/__tests__/compute-economics-real-math.test.ts
- move EconomicsInput interface to lib/types.ts (was only a type import in pipeline.ts)
- update lib/pipeline.ts import source
- mark stale LOW-1 finding in .pipeline/phase_3_qi.md
- add Wave 2 deletion note in docs/audits/2026-05-28-test-coverage-audit.md

Tests: 7 economics suites (59 tests) + 8 pipeline suites (46 tests) green.
tsc --noEmit: clean.
