+++ docs/audits/2026-05-28-test-coverage-audit.md
@@ -296,3 +296,5 @@ Across ~813 test files the suite is **broad but dishonest in its most security- 
 - **N-lib-core-2**: The lib-core-2 domain is generally in very good health: ROI metrics, session management, subs-guardian, corpus/MIME, plan-first, and rate-limiter tests are thorough and exercise real business logic with meaningful edge cases. The three real issues are isolated: a structurally tautological AuditEntry shape test that provides no type-safety regression value, three trivially-satisfied magnitude assertions in sof-parser that would hide a completely broken parser, and a real-timer TTL test with a 15ms margin that is a latent source of CI flakiness. The non-LLMTimeoutError re-throw path in voice-transcribe is the only meaningful coverage gap worth adding a test for.
 - **O-other-1**: The unit/integration tests (demo-mode-validator, l5c-extras, l5c-hierarchy, subs-guardian, cron-init-db) are well-structured and exercise real SUT logic against in-memory databases. The main systemic gap is that deadline/cron tests all use live Date.now() to place deadlines near stage boundaries rather than injecting a fixed clock, making the stage-selection logic implicitly time-dependent. The E2E Playwright suite is large but appropriately guarded with feature-gate skips; the only substantive issue is the broker-happy-path second test which is tautologically weak.
 - **O-other-2**: The mixed-domain tests in this chunk are of generally good quality: pure-function tests (vessels-utils, openai-helpers, ports/resolve, format-number, format-date, event-emitter, process-email) are behavioural and meaningful. The main systemic gap is a recurring pattern where tests verify regression fixes via static JSX/source-file string-matching (557-laycan, mode-aware-content) rather than functional component or route-handler invocation — these provide weak confidence because they test code structure rather than runtime behaviour. Two functions (extractSpecNumbers, validateExplainDealResponse) are imported but never called in the regression test, masking a real coverage hole in the validator module.
+
+**Wave 2 update (2026-06-10):** `computeEconomics` in `lib/economics/index.ts` and its tests (`__tests__/index.test.ts`, `__tests__/compute-economics-real-math.test.ts`) deleted as dead code — findings 5 and N-lib-core-1 regarding this function are no longer applicable.
